move large media card styles to be on their own for easier editing
